Update views for integer volgnummer

+++ gobcore/views/bag/verblijfsobjecten/enhanced.sql
@@ -80,19 +80,19 @@
       bag_nummeraanduidingen AS heeft_hoofdadres
     ON
       vot.heeft_hoofdadres->>'id' = heeft_hoofdadres._id AND
-      vot.heeft_hoofdadres->>'volgnummer' = heeft_hoofdadres.volgnummer
+      (vot.heeft_hoofdadres->>'volgnummer')::int = heeft_hoofdadres.volgnummer
     -- SELECT ligt_aan_openbareruimte
     LEFT JOIN
       bag_openbareruimtes AS ligt_aan_openbareruimte
     ON
       heeft_hoofdadres.ligt_aan_openbareruimte->>'id' = ligt_aan_openbareruimte._id AND
-      heeft_hoofdadres.ligt_aan_openbareruimte->>'volgnummer' = ligt_aan_openbareruimte.volgnummer
+      (heeft_hoofdadres.ligt_aan_openbareruimte->>'volgnummer')::int = ligt_aan_openbareruimte.volgnummer
     -- SELECT ligt_in_woonplaats
     LEFT JOIN
       bag_woonplaatsen AS ligt_in_woonplaats
     ON
       heeft_hoofdadres.ligt_in_woonplaats->>'id' = ligt_in_woonplaats._id AND
-      heeft_hoofdadres.ligt_in_woonplaats->>'volgnummer' = ligt_in_woonplaats.volgnummer
+      (heeft_hoofdadres.ligt_in_woonplaats->>'volgnummer')::int = ligt_in_woonplaats.volgnummer
     -- SELECT ligt_in_gemeente
     LEFT JOIN
       brk_gemeentes AS ligt_in_gemeente
@@ -125,42 +125,42 @@
       bag_panden AS pand_bouwblok
     ON
       vot.ligt_in_panden#>>'{0,id}' = pand_bouwblok.identificatie AND
-      vot.ligt_in_panden#>>'{0,volgnummer}' = pand_bouwblok.volgnummer
+      (vot.ligt_in_panden#>>'{0,volgnummer}')::int = pand_bouwblok.volgnummer
     LEFT JOIN
       gebieden_bouwblokken AS ligt_in_bouwblok
     ON
       pand_bouwblok.ligt_in_bouwblok->>'id' = ligt_in_bouwblok.identificatie AND
-      pand_bouwblok.ligt_in_bouwblok->>'volgnummer' = ligt_in_bouwblok.volgnummer
+      (pand_bouwblok.ligt_in_bouwblok->>'volgnummer')::int = ligt_in_bouwblok.volgnummer
     -- SELECT ligt_in_buurt
     LEFT JOIN
       gebieden_buurten AS ligt_in_buurt
     ON
       vot.ligt_in_buurt->>'id' = ligt_in_buurt.identificatie AND
-      vot.ligt_in_buurt->>'volgnummer' = ligt_in_buurt.volgnummer
+      (vot.ligt_in_buurt->>'volgnummer')::int = ligt_in_buurt.volgnummer
     -- SELECT ligt_in_wijk
     LEFT JOIN
       gebieden_wijken AS ligt_in_wijk
     ON
       ligt_in_buurt.ligt_in_wijk->>'id' = ligt_in_wijk.identificatie AND
-      ligt_in_buurt.ligt_in_wijk->>'volgnummer' = ligt_in_wijk.volgnummer
+      (ligt_in_buurt.ligt_in_wijk->>'volgnummer')::int = ligt_in_wijk.volgnummer
     -- SELECT ligt_in_stadsdeel
     LEFT JOIN
       gebieden_stadsdelen AS ligt_in_stadsdeel
     ON
       ligt_in_wijk.ligt_in_stadsdeel->>'id' = ligt_in_stadsdeel.identificatie AND
-      ligt_in_wijk.ligt_in_stadsdeel->>'volgnummer' = ligt_in_stadsdeel.volgnummer
+      (ligt_in_wijk.ligt_in_stadsdeel->>'volgnummer')::int = ligt_in_stadsdeel.volgnummer
     -- SELECT _ligt_in_ggwgebied
     LEFT JOIN
       gebieden_ggwgebieden AS ligt_in_ggwgebied
     ON
       ligt_in_buurt._ligt_in_ggwgebied->>'id' = ligt_in_ggwgebied.identificatie AND
-      ligt_in_buurt._ligt_in_ggwgebied->>'volgnummer' = ligt_in_ggwgebied.volgnummer
+      (ligt_in_buurt._ligt_in_ggwgebied->>'volgnummer')::int = ligt_in_ggwgebied.volgnummer
     -- SELECT _ligt_in_ggpgebied
     LEFT JOIN
       gebieden_ggpgebieden AS ligt_in_ggpgebied
     ON
       ligt_in_buurt._ligt_in_ggpgebied->>'id' = ligt_in_ggpgebied.identificatie AND
-      ligt_in_buurt._ligt_in_ggpgebied->>'volgnummer' = ligt_in_ggpgebied.volgnummer
+      (ligt_in_buurt._ligt_in_ggpgebied->>'volgnummer')::int = ligt_in_ggpgebied.volgnummer
     WHERE
       (vot._expiration_date > current_date OR vot._expiration_date IS NULL)
       AND vot._date_deleted IS NULL
